% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{antgreens_opt_name}
\alias{antgreens_opt_name}
\title{Variable Name of Antgreens Options List}
\format{
An object of class \code{character} of length 1.
}
\usage{
antgreens_opt_name
}
\description{
Variable Name of Antgreens Options List
}
\keyword{datasets}
