% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_report.R
\name{auto_report}
\alias{auto_report}
\title{Automatically Render Report}
\usage{
auto_report(input_dir = ".", data_path, skip_conv = FALSE, ...)
}
\arguments{
\item{input_dir}{character, default '.'. Root directory for a bookdown project
(contains \code{index.Rmd} file).}

\item{data_path}{void}

\item{skip_conv}{void}

\item{...}{void}
}
\value{
path of HTML book.
}
\description{
Automatically Render Report
}
