% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{bar_line_echart}
\alias{bar_line_echart}
\title{Bar and Line Chart (with Timeline)}
\usage{
bar_line_echart(
  data,
  timeline_var = NULL,
  x_var,
  bar_var,
  line_var,
  main_axis_name,
  sec_axis_name,
  x_axis_name,
  axis_fontsize = 12,
  x_name_gap = 25,
  long_x_label = FALSE,
  x_label_width = 30,
  use_y_upper_bound = FALSE,
  show_label = TRUE,
  toolbox_features = c("saveAsImage", "restore", "dataView")
)
}
\arguments{
\item{data}{a data frame}

\item{timeline_var}{character. The variable indicating timeline.}

\item{x_var}{character. The variable that will be the X axis in the chart.}

\item{bar_var}{character vector. The variable that will be used to draw bar chart.}

\item{line_var}{character vector. The variable that will be used to draw line chart.}

\item{main_axis_name}{character. The name of main Y axis. If missing, the function will use
information in \code{option} to auto-detect the name according to \code{bar_var}.}

\item{sec_axis_name}{character. The name of secondary Y axis. If missing, the function will use
information in \code{option} to auto-detect the name according to \code{line_var}.}

\item{x_axis_name}{character. The name of X axis. If missing, the function will use
information in \code{option} to auto-detect the name according to \code{x_var}.}

\item{axis_fontsize}{numeric, default 12. The font size of axis name.}

\item{x_name_gap}{numeric, default 25. The distance between x axis and label of x axis.}

\item{long_x_label}{logical, default FALSE. Whether to use style that solving the problem than
the label text of x axis is too long.}

\item{x_label_width}{numeric, default 30. The label text width of x axis.}

\item{use_y_upper_bound}{logical, default FALSE. Whether to use upper bound of Y axis.}

\item{show_label}{logical, default TRUE. Whether to show data labels.}

\item{toolbox_features}{character vectors, default \code{c('saveAsImage', 'restore', 'dataView')}.
It controls the toolbox functions in echarts.}
}
\value{
an object of class \code{echarts4r}.
}
\description{
Plot bar and line chart using \code{\link[echarts4r]{e_bar}} and
\code{\link[echarts4r]{e_line}}. Regardless of whether there is a timeline axis,
this function can plot charts correctly.
}
