% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{bar_line_ggplot}
\alias{bar_line_ggplot}
\title{Plot bar and line figure in the same figure}
\usage{
bar_line_ggplot(
  dataset,
  x_var,
  bar_var = "sample_size",
  scale_blank_div = 5,
  scale_interval = NA,
  main_axis_name,
  sec_axis_name,
  x_axis_name,
  bar_fill_palette = "grey",
  line_color = "black",
  point_color = "white",
  legend_pos = c("none", "left", "right", "bottom", "top"),
  label_size = 12
)
}
\arguments{
\item{label_size}{}
}
\value{

}
\description{
Plot bar and line figure in the same figure
}
