% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_report.R
\name{book_env}
\alias{book_env}
\title{Generate Environment for Book Rendering}
\usage{
book_env(data_path, skip_conv = FALSE, year_range = NA, year_colname = "年份")
}
\arguments{
\item{data_path}{character. Path of csv or excel raw data file}

\item{skip_conv}{logical, default FALSE. If set TRUE then the \code{data_path}
refers to the path of CSV raw data file, otherwise refers to the path of .xls(x)
raw data file.}

\item{year_range}{numeric vector length 2, default NA.}

\item{year_colname}{character of column name for dimension \code{year}.}
}
\value{
environment object contains necessary objects.
}
\description{
Generate Environment for Book Rendering
}
