% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_residue_index}
\alias{calc_residue_index}
\title{Calculate residue index using MRL list}
\usage{
calc_residue_index(vm, products, mrl_map)
}
\arguments{
\item{vm}{test values matrix of each product (row) in each drug (column)}

\item{products}{character vector composed of product names, which correspond to each
row in param `vm`}

\item{mrl_map}{a R object List that the product name is used as a key and the MRL
standards vector on pesticides is used as a value.}
}
\value{
numeric matrix where each value have to be divided by corresponding MRL standard.
}
\description{
this Rcpp function `calc_use_mrl()` receives test values matrix and
returns a numeric matrix that each value have to be divide by correspoding MRL value.
}
