% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_treated_data.R
\name{calc_treated_data}
\alias{calc_treated_data}
\title{Calculate treated data using raw data and MRL}
\usage{
calc_treated_data(
  raw_data,
  mrl_data,
  st_drug,
  ed_drug,
  raw_col_en2cn = RawColnameTrans,
  product_name = ProductLabelName
)
}
\arguments{
\item{raw_data}{a data frame. It is raw detection data of each sample on each drug.
It has multiple columns as labels of each sample and the remaining columns are
raw detection values.}

\item{mrl_data}{a data frame. It has one column named `产品名称` and the other columns
named with name of drugs. Every row is MRL value of one product on each drugs.}

\item{st_drug}{a character. Name of the first column of drug.}

\item{ed_drug}{a character. Name of the last column of drug.}

\item{raw_col_en2cn}{list, default \code{\link{RawColnameTrans}}. It is the
English to Chinese names of dimensions except product.}

\item{product_name}{character. It is the product dimension name.}
}
\value{
a data frame with 3 extra columns compared with `raw_data`.
}
\description{
`calc_treated_data()` construct a new data frame which construction
is like raw_data that has multiple columns as labels and the remaning columns are
detection values of each drug on each sample.
}
\examples{
# library(antgreens)
# set_dimensions("name", list(t1="t1", t2="t2"))
# df <- data.frame(name = c('a', 'b', 'a', 'c', 'c'),t1 = c(1,2,3,4,5),t2 = c(0,9,8,7,6))
# mrl <- data.frame(name = c('a','b','c'),t1 = c(2,2,4),t2 = c(4,10,6))
# mrl_map <- mapping_mrl(as.matrix(mrl[,2:3]), mrl$product)
# nvm <- calc_treated_data(as.matrix(df[,2:3]), mrl_map, "t1", "t2")
}
