% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{change_to_defective}
\alias{change_to_defective}
\title{Change Qulification Rate (Percent) to Fractration Defective (Percent)}
\usage{
change_to_defective(data, use_percent = TRUE, digits = 2)
}
\arguments{
\item{data}{data frame with \code{qualification_rate(_percent)} column}

\item{use_percent}{logical, default TRUE. Whether to use "percent" suffix.}

\item{digits}{numeric, default 2. The number of dicimal reserved when using minus to change
to defective.}
}
\value{
a data frame with \code{fractration_rate(_percent)} column.
}
\description{
\code{change_to_defective(df)} will change the \code{qualification_rate} column in
\code{df} to \code{defective_rate} using \eqn{1 - qualification_rate}, or change the
\code{qualification_rate_percent} column in \code{df} to \code{defective_rate_percent} column
using \eqn{100 - qualification_rate_percent}.
}
\examples{
df <- data.frame(qualification_rate = runif(3, 0, 1), qualification_rate_percent = runif(3, 0, 100))
change_to_defective(df, use_percent = FALSE)
change_to_defective(df)
}
