% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_required_pkgs.R
\name{check_required_pkgs}
\alias{check_required_pkgs}
\title{Check the required packages and install them}
\usage{
check_required_pkgs(pkgs_vec = c("data.table", "readxl", "readr", "bookdown"))
}
\arguments{
\item{pkgs_vec}{a character vector contains names of required packages.}
}
\value{
None (invisible `NULL`)
}
\description{
Check the required packages and install them
}
\examples{
# library(antgreens)
# pkgs <- c('data.table', 'tidyverse', 'Rcpp')
# check_required_pkgs(pkgs)
}
