% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sample_size.R
\name{check_sample_size}
\alias{check_sample_size}
\title{Check Sample Size of combinations of dimensions}
\usage{
check_sample_size(data, dimensions, drug_names = DrugNames, path)
}
\arguments{
\item{data}{data frame of raw data or renamed data.}

\item{dimensions}{character vector of dimensions to be analysis in \code{data}.}

\item{drug_names}{dimension name for \code{drug}. The default value
is in \code{\link{DrugNames}}.}

\item{path}{character of \code{.xls(x)} file path.}
}
\value{
a data frame of dimension combination which are encoded with 0/1 and
maximum sample size (\code{max_ss}), median sample size (\code{med_ss}) and minimum
sample size (\code{min_ss}).
}
\description{
Check Sample Size of combinations of dimensions
}
\examples{
data <- data.frame(a=c(1,2,1,2,1,2,1,2), b=c(1,1,2,2,3,3,1,1), c=c(1,1,1,1,1,1,2,2))
check_sample_size(data, dimensions = c('a', 'b', 'c'))
}
