% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingustic.R
\docType{data}
\name{ColnamesI18n}
\alias{ColnamesI18n}
\title{I18N Settings for Displaying Column Names}
\format{
An object of class \code{list} of length 1.
}
\usage{
ColnamesI18n
}
\description{
I18N Settings for Displaying Column Names
}
\keyword{datasets}
