% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{combine_year_quarter}
\alias{combine_year_quarter}
\title{Combine the Year and Quarter Column into '2021-Q1' Form}
\usage{
combine_year_quarter(data, year_name = "year", quarter_name = "quarter")
}
\arguments{
\item{data}{a data frame.}

\item{year_name}{character, default 'year'. It is the column name representing the year.}

\item{quarter_name}{character, default 'quarter'. It is the column name representing the quarter.}
}
\value{
a data frame with column \code{timeline} and without column \code{year_name} and
\code{quarter_name}.
}
\description{
This function will combine the numeric or factor column of year and quarter into
the form of "2021-Q1" whcih reffers to the year 2021 and the first quarter.
}
\examples{
df <- data.frame(a = 2017:2021, b = c(1,2,2,4,3))
combine_year_quarter(df, year_name = 'a', quarter_name = 'b')
}
