% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingustic.R
\docType{data}
\name{DataTableI18n}
\alias{DataTableI18n}
\title{I18N Settings for DataTable}
\format{
An object of class \code{list} of length 1.
}
\usage{
DataTableI18n
}
\description{
I18N Settings for DataTable
}
\keyword{datasets}
