% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sample_size.R
\docType{data}
\name{DrugName}
\alias{DrugName}
\title{Names of "Drug"}
\format{
An object of class \code{character} of length 4.
}
\usage{
DrugName
}
\description{
\code{DrugName} is a character vector with some names of "drug"
using in vegetables.
}
\keyword{datasets}
