% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingustic.R
\docType{data}
\name{DrugNames}
\alias{DrugNames}
\title{Names of "Drug"}
\format{
An object of class \code{character} of length 2.
}
\usage{
DrugNames
}
\description{
\code{DrugName} is a character vector with some names of "drug"
using in vegetables.
}
\keyword{datasets}
