% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{DTOptions}
\alias{DTOptions}
\title{Options for DT Package}
\format{
An object of class \code{list} of length 8.
}
\usage{
DTOptions
}
\description{
Options for DT Package
}
\keyword{datasets}
