% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_utils.R
\name{factor_dims}
\alias{factor_dims}
\title{Make Columns As Factor}
\usage{
factor_dims(
  dataset,
  ...,
  funcs = list(),
  quarter_readable = TRUE,
  quarter_trans = antgreens::quarter_readable
)
}
\arguments{
\item{dataset}{data frame with readable \code{quarter} column
or \code{year}/\code{province}/\code{category}/\code{product}/\code{drug}/\code{timeline} column.}

\item{...}{dimensions to be made as factor.}

\item{funcs}{named list of custome factor function}

\item{quarter_readable}{logical, default \code{TRUE}. If set \code{TRUE} then make
quarter column readable and as a factor.}

\item{quarter_trans}{list containing 4 characters and the names of each element is 1, 2, 3, 4,
which represents the four quarters.}
}
\value{
data frame with factor column. And the order of quarter is
depends on the order of \code{quarter_readable}; the order of year is depends on
the number of year; the order of province, category product and drug are depends
on the dictionary order of character.
}
\description{
Make Columns As Factor
}
\examples{
data <- data.frame(year = 2016:2019, quarter = c(1,2,3,4))
factor_dims(data, year, quarter, quarter_trans = list(`1`='a',`2`='b',`3`='c',`4`='d'))
}
