% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_utils.R
\name{factor_year}
\alias{factor_year}
\alias{factor_quarter}
\alias{factor_province}
\alias{factor_category}
\alias{factor_product}
\alias{factor_drug}
\alias{factor_timeline}
\title{Make Column As Factor}
\usage{
factor_year(dataset)

factor_quarter(
  dataset,
  readable = FALSE,
  quarter_trans = antgreens::quarter_readable
)

factor_province(dataset)

factor_category(dataset)

factor_product(dataset)

factor_drug(dataset)

factor_timeline(dataset)
}
\arguments{
\item{dataset}{data frame}

\item{readable}{logical, default \code{TRUE}. If set \code{TRUE} then make
quarter column readable and as a factor.}

\item{quarter_trans}{list containing 4 characters and the names of each element is 1, 2, 3, 4,
which represents the four quarters.}
}
\description{
The subject variants of `factor_*` has:
  * `year`: factor numeric year variable
  * `quarter`: (make quarter readable and) factor quarter character
  * `province`: factor province character
  * `category`: factor category character
  * `product`: factor product character
  * `drug`: factor drug character
  * `timeline`: factor timeline character
}
\examples{
data <- data.frame(
year = 2016:2019, quarter = c(1,2,3,4), province = c('ab','bc','cd','de'), category = c('ab','bc','cd','de'), product = c('ab','bc','cd','de'), drug = c('ab','bc','cd','de'))
quarter_trans <- list(`1`='a',`2`='b',`3`='c',`4`='d')
# factor year
factor_year(data)
# factor quarter
factor_quarter(data)
factor_quarter(data, readable = TRUE, quarter_trans = quarter_trans)
# factor province
factor_province(data)
# factor category
factor_category(data)
# factor product
factor_product(data)
# factor drug
factor_drug(data)
# factor timeline
factor_timeline(combine_year_quarter(data))
}
