% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_multi_dims_stats.R
\name{gen_dims_comb_stats}
\alias{gen_dims_comb_stats}
\title{Generate stats for all combinations of dims to be analysis}
\usage{
gen_dims_comb_stats(data, all_dims, fast = TRUE)
}
\arguments{
\item{data}{data.table if `fast` is `TRUE` where data.frame if `fast` is FALSE. It is
the data divided by MRL so you should evaluate function `calc_treated_data()` before
using it.}

\item{all_dims}{character vector of all dimensions.}

\item{fast}{logic, default is `TRUE`. Determine whether to use a more efficient function
rewrite with data.table.}
}
\value{
a list contains $2^length(all_dims)$ data.frame. Each data.frame is named with
dimensions' names connected with underline and contains dimensions columns and
statistic columns.
}
\description{
`gen_dims_comb_stats()` returns a list contains $2^length(all_dims)$
data frame. It will calculate the specified statistical indicators for each combination
of dimensions to be analysis.
}
\examples{
# treated_data <- calc_treated_data(raw_data, mrl_data, "\u7532\u80fa\u78f7", "\u4e8c\u7532\u620a\u7075")
# dims <- c(names(RawColumnTrans), "product", "drug")
## fast mode
# gen_dims_comb_stats(data, dims)
## slow mode
# gen_dims_comb_stats(treated_data, dims, fast = FALSE)
}
