% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{get_bounds}
\alias{get_bounds}
\title{Get Lower/Upper Bounds of Numeric/Integer Sequence}
\usage{
get_bounds(
  data,
  column_name,
  increment = 5,
  upper_shrink = FALSE,
  shrink_prop = 2/3
)
}
\arguments{
\item{data}{a data frame with numeric or integer sequence.}

\item{column_name}{character of column name.}

\item{increment}{numeric, default 5. It is the increment of bounds.}

\item{upper_shrink}{locical, default FALSE. If set TRUE and the \code{shrink_prop}
of upper bound is larger than the maximum of target column, the actual upper column
will use its maximum value.}

\item{shrink_prop}{numeric, default is 0.66666.... This is the proportion of shrink threshold.}
}
\value{
numeric vector of length 2. The first is lower bound and the second is upper bound.
}
\description{
Get Lower/Upper Bounds of Numeric/Integer Sequence
}
\examples{
df <- data.frame(a = c(11, 23, 1), b = c(0.2, 0.3, 0.1))
get_bounds(df, 'a')
get_bounds(df, 'b', increment = 0.2)
get_bounds(df, 'b', increment = 0.4, upper_shrink = TRUE)
}
