% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_treated_data.R
\name{get_multi_defective_num}
\alias{get_multi_defective_num}
\title{Get multiple defective number}
\usage{
get_multi_defective_num(res_ind_vector)
}
\arguments{
\item{res_ind_vector}{It can be a numeric vector or
data.frame with one numeric row. And its significance should
be the ratio of the detection value of samples on each
pesticide to MRL(maximum residue limit).}
}
\value{
An integer number that is multiple defective number.
}
\description{
This function calculate the multiple
defective number of one sample. The value of multiple
defective number is the number of components greater
than 1 in a vector.
}
\examples{
# when res_ind_vector is a numeric vector:
x <- c(0.1, 0.2, 0, 0, 0.3)
get_multi_defective_num(x)

# when want to apply this function on a data.frame:
library(dplyr)
library(magrittr)
df <- data.frame(a=c(0.1,0,0.3), b=c(0,0,0), c=c(0,0.2,0.3))
df \%>\%
    rowwise() \%>\%
    mutate(mdn = get_multi_defective_num(c_across(a:c)))
}
