% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{get_odds_bounds}
\alias{get_odds_bounds}
\title{Get Lower/Upper Bounds for Odds to Display}
\usage{
get_odds_bounds(data, column_name, lower = TRUE, increment = 5)
}
\arguments{
\item{data}{a data frame.}

\item{column_name}{character. Name of target column which value is between [0, 1] or [0, 100].}

\item{lower}{logical, default TRUE. If set TRUE then will return the lower bound, otherwise
will return the upper bound.}

\item{increment}{numeric, default is 5. It is the increment of bound.}
}
\value{
numeric
}
\description{
Get Lower/Upper Bounds for Odds to Display
}
\examples{
df <- data.frame(a = c(11, 23, 1), b = c(0.2, 0.3, 0.1))
get_odds_bounds(df, 'a', lower = FALSE)
get_odds_bounds(df, 'b', increment = 0.2, lower = FALSE)
}
