% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_option}
\alias{get_option}
\title{Get \code{antgreens_options} in \code{.GlobalEnv}}
\usage{
get_option(var_name)
}
\arguments{
\item{var_name}{character. The key of variable \code{antgreens_opt_name}.}
}
\description{
Get \code{antgreens_options} in \code{.GlobalEnv}
}
