% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sample_size.R
\name{get_sample_size_stats}
\alias{get_sample_size_stats}
\title{Get Sample Size Statistics for Given Dimension Combnation}
\usage{
get_sample_size_stats(dataset, draw_figure = TRUE, ...)
}
\arguments{
\item{dataset}{data frame of raw data or renamed data.}

\item{draw_figure}{logical, default \code{TRUE}. If \code{TRUE} this function will
draw histogram of sample size of given dimension combination.}

\item{...}{dimensions}
}
\value{
a list with sample size data frame, minimum, median and maximum sample
size.
}
\description{
Get Sample Size Statistics for Given Dimension Combnation
}
\examples{
data <- data.frame(a=c(1,2,1,2,1,2,1,2), b=c(1,1,2,2,3,3,1,1), c=c(1,1,1,1,1,1,2,2))
get_sample_size_stats(data, draw+figure = FALSE, a, b)
}
