% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{get_threshold}
\alias{get_threshold}
\title{Get Corresponding Threshold Value}
\usage{
get_threshold(threshold_df, dims, arg_dims)
}
\arguments{
\item{threshold_df}{data frame contains dimension labels encoded with 0/1 and
threshold values.}

\item{dims}{character vector of all dimension variable names.}

\item{arg_dims}{character vector of selected dimension variable names.}
}
\value{
nunmeric value
}
\description{
Get Corresponding Threshold Value
}
