% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_report.R
\name{get_year_range}
\alias{get_year_range}
\title{Get the year span of the analysis}
\usage{
get_year_range(csv_data_file, year_colname = "年份")
}
\arguments{
\item{csv_data_file}{a character representing the UTF-8 `.csv` data file.}

\item{year_colname}{a character. It is the name of column that represents the year.}
}
\value{
a numeric vector of length 2, representing min and max year separately.
}
\description{
Get the year span of the analysis
}
