% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{heatmap_echart}
\alias{heatmap_echart}
\title{Heatmap Chart (with Timeline)}
\usage{
heatmap_echart(
  data,
  timeline_var = NULL,
  x_var,
  y_var,
  value_var,
  x_axis_name,
  y_axis_name,
  value_axis_name,
  width = "100\%",
  height = "700\%",
  shrink_height = NULL,
  x_label_width = 1,
  x_label_fontsize = 12,
  toolbox_freatures = c("saveAsImage", "restore")
)
}
\arguments{
\item{data}{a data frame}

\item{timeline_var}{character. The variable indicating timeline.}

\item{x_var}{character. The variable that will be the X axis in the chart.}

\item{y_var}{character. The variable that will be the Y axis in the chart.}

\item{value_var}{character. The variable that will be displayed in cells in the chart.}

\item{x_axis_name}{character. The name of X axis. If missing, the function will use
information in \code{option} to auto-detect the name according to \code{x_var}.}

\item{y_axis_name}{character. The name of Y axis. If missing, the function will use
information in \code{option} to auto-detect the name according to \code{y_var}.}

\item{value_axis_name}{character. The name of value variable. If missing, the function will use
information in \code{option} to auto-detect the name according to \code{value_var}.}

\item{width, height}{character, \code{width} default is "100%" and \code{height} default is
"700%". Must be a valid CSS unit (like \code{'100\%'}, \code{'400px'}, \code{'auto'})
or a number, which will be coerced to a string and have \code{'px'} appended.}

\item{shrink_height}{character, default NULL. It is the same as \code{width} and \code{height}
but to used in \code{\link[echarts4r]{e_grid}} to control the height of figure grid so the
long label in X axis can be seen fully.}

\item{x_label_width}{numeric, default 1. The width of text of x label. If the text width overflow,
then will use break to wrap.}

\item{x_label_fontsize}{numeric, default 12. The font size of text of x label.}

\item{toolbox_freatures}{character vectors, default \code{c("saveAsImage", "restore")}. It controls
the toolbox functions in echarts.}
}
\value{
an object of class \code{echarts4r}.
}
\description{
Plot heat map using \code{\link[echarts4r]{e_heatmap}}. This function uses
\code{datazoom} toolbox with type "inside" and "continuous" visual map. For Y axis labels,
the width of label text is set to 60 and when it will use break to wrap when the actual
width is overflow. For timeline options, the timeline widget is on the top and 5 units from
top. The time point widgets padding is set to 0 and the play intercal is 1500 ms.
}
