% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_treated_data.R
\name{is_qualified}
\alias{is_qualified}
\title{Get qualified label}
\usage{
is_qualified(res_ind_vector)
}
\arguments{
\item{res_ind_vector}{It can be a numeric vector or
data.frame with one numeric row. And its significance should
be the ratio of the detection value of samples on each
pesticide to MRL(maximum residue limit).}
}
\value{
If a sample is calculating to be qualified, return 1. Otherwise return 0.
}
\description{
This function check whether there exists at least
one component that greater than 1 in a vector. If all components are
less than or equal to 1, this sample is qualified, otherwise is not qualified.
}
\examples{
# qualified sample (vector)
q <- c(0,0,0,0,0.01,0,0,0.2)
is_qualified(q)

# unqualified sample (vector)
uq <- c(0,0,0,0,0.01,0,0,1.2)
is_qualified(uq)

# samples (data.frame)
library(dplyr)
library(magrittr)
df <- data.frame(a=c(0.1,0,0.3), b=c(0,1,0), c=c(0,0.2,0.3))
df \%>\%
    rowwise() \%>\%
    mutate(is_qualified = is_qualified(c_across(a:c)))
}
