% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{load_antgreens_options}
\alias{load_antgreens_options}
\title{Load Initial Options for antgreens}
\usage{
load_antgreens_options()
}
\value{
None
}
\description{
Load Initial Options for antgreens
}
\examples{
# load_antgreens_options()
}
