% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_threshold.R
\name{load_threshold}
\alias{load_threshold}
\title{Load Threshold Data in Excel Format}
\usage{
load_threshold(path)
}
\arguments{
\item{path}{character of .xls(x) file which has threshold data.}
}
\value{
a data frame contains dimensions combination encoded with 0/1 and threshold
data for each combination.
}
\description{
Load Threshold Data in Excel Format
}
