% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{make_quarter_readable}
\alias{make_quarter_readable}
\title{Make Quarter Readable in Table}
\usage{
make_quarter_readable(df)
}
\arguments{
\item{df}{data frame contains column named with "quarter"}
}
\value{
data frame with readable "quarter" column
}
\description{
use readable name of quarters to replace number of quarters.
}
\examples{
df <- data.frame(quarter = c(1,2,3,4))
df <- make_quarter_readable(df)
}
