% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{map_echart}
\alias{map_echart}
\title{Map Chart (with Timeline)}
\usage{
map_echart(
  data,
  map_json,
  timeline_var = NULL,
  value_var = "defective_rate_percent",
  value_axis_name,
  use_digits = TRUE,
  interval = 0,
  timeline_label_width = 50,
  toolbox_features = c("dataView", "saveAsImage", "restore")
)
}
\arguments{
\item{data}{a data frame}

\item{map_json}{A geo list containing map data}

\item{timeline_var}{character. The variable indicating timeline.}

\item{value_var}{character. The variable that will be displayed in cells in the chart.}

\item{value_axis_name}{character. The name of value variable. If missing, the function will use
information in \code{option} to auto-detect the name according to \code{value_var}.}

\item{use_digits}{logical, default TRUE. Whether to use digits in formatter.}

\item{interval}{numeric, default 0. The interval of label of timeline points.}

\item{timeline_label_width}{numeric, default 50. The label text width of timeline points.}

\item{toolbox_features}{character vectors, default \code{c("dataView", "saveAsImage", "restore")}.
It controls the toolbox functions in echarts.}
}
\value{
an object of class \code{echarts4r}.
}
\description{
Plot map using \code{\link[echarts4r]{e_map}}. It will auto-detect the \code{value_var}
name using information in \code{option} when the \code{value_axis_name} is not given. The chart's
width is set "100\%" and the height is set "800\%" to make the map larger enougth to see. For visual
map, the type is "continuous" and the upper bound is calculated automatically. The position of
visual map is set to right and vertically centered.
}
