% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mapping_mrl}
\alias{mapping_mrl}
\title{Mapping MRL data frame by line}
\usage{
mapping_mrl(mrl, products)
}
\arguments{
\item{mrl}{MRL numeric matrix of each product (row) in each drug (column)}

\item{products}{character vector composed of product names, which correspond to each
row in param `mrl`}
}
\value{
std::map<String, NumericVector> in Rcpp and List in R.
}
\description{
this Rcpp function `mapping_mrl()` construct a map object
<String, NumericVector> using prduct name as key and numbers of drugs' MRL values
vector of this product as value. In this case, users can easily and efficiently query
MRL numeric vector with product's name.
}
