% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_multi_dims_stats.R
\name{multi_dims_stats}
\alias{multi_dims_stats}
\title{Get multi-dimensions stats}
\usage{
multi_dims_stats(
  data,
  ...,
  drug_colname = "drug",
  st_drug_name = "甲胺磷",
  ed_drug_name = "二甲戊灵"
)
}
\arguments{
\item{data}{a data frame with dimensions labels and drug monitoring data divided by MRL.}

\item{...}{dimension names parameters. **NOT Character**. The supported dimensions are
depending on `data`. As for now, only 6 dimensions are supported:

  * year
  * quarter
  * province
  * category
  * product
  * drug (depending on param `drug_colname`)

If you want to use (and you must use) this param, you should replace `...`
with e.g. `quarter, province, drug`.}

\item{drug_colname}{the name of column containing drug names, and is the dimension name.}

\item{st_drug_name}{the name of the first drug residual column.}

\item{ed_drug_name}{the name of the last drug residual column. *Notice*: the column
between `st_drug_name` and `ed_drug_name` should all be drug residual column.}
}
\value{
a data frame containing essential stats.
}
\description{
`multi_dims_stats()` creates a new data frame. It will have one or more
rows for each combination of grouping dimensions. If there are no grouping dimensions,
the output will have a single row summarising all samples in the input data frame.
Despite the dimensions columns, it will contain other 3 or 6 columns depending on
whether there is *drug* dimension in the dimension combination.

If *drug* dimension is in the dimension combination, it will report:

  * `sample_size`: is the size of the sample set under specified dimensions.
  * `detection_rate`: is the detection rate of this specified sample set.
  * `qualification_rate`: is the qualification rate of this specified sample set.

If *drug* dimension is **not** in the dimension combination, it will report:

  * `sample_size`: is the size of the sample set under specified dimensions.
  * `qualification_rate`: is the qualification rate of this specified sample set.
  * `multi_detection_rate`: is the multi-residue detection rate of this specified
  sample set.
  * `max_defective_num`: is the max multi-residue detection number of this
  specified sample set.
  * `multi_defective_rate`: is the multi-residue defective rate of this specified sample set.
  * `max_defective_num`: is the multi-residue defective number of this
  specified sample set.
}
\examples{
## without `drug`
# multi_dims_stats(treated_data, year, province)
# multi_dims_stats(treated_data)
## with `drug`
# multi_dims_stats(treated_data, quarter, drug)
}
