% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{normalize_path}
\alias{normalize_path}
\title{Normalize paths}
\usage{
normalize_path(path, winslash = .Platform$file.sep, must_work = FALSE)
}
\arguments{
\item{path}{character or character vector of file paths.}

\item{winslash}{the separator to be used on Windows. Default is
\code{\link{.Platform}$file.sep}. Must be one of \code{c('/', '\\\\')}.}

\item{must_work}{logical, default FALSE.}
}
\description{
a wrapper function of \code{\link{normalizePath}()} with different
default parameters.
}
\examples{
library(antgreens)
normalize_path(c('', NA, "file/path"))
}
