% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingustic.R
\docType{data}
\name{ProductLabelName}
\alias{ProductLabelName}
\title{Label Name of Product in Raw and MRL Data}
\format{
An object of class \code{character} of length 1.
}
\usage{
ProductLabelName
}
\description{
Label Name of Product in Raw and MRL Data
}
\keyword{datasets}
