% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingustic.R
\docType{data}
\name{ProvincePinyin}
\alias{ProvincePinyin}
\title{PinYin of Provinces in China}
\format{
An object of class \code{character} of length 34.
}
\usage{
ProvincePinyin
}
\description{
PinYin of Provinces in China
}
\keyword{datasets}
