% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingustic.R
\docType{data}
\name{ProvinceShort}
\alias{ProvinceShort}
\title{Short Names of Provinces in China}
\format{
An object of class \code{character} of length 34.
}
\usage{
ProvinceShort
}
\description{
Short Names of Provinces in China
}
\keyword{datasets}
