% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingustic.R
\docType{data}
\name{quarter_readable}
\alias{quarter_readable}
\title{Quarters Readable Name}
\format{
An object of class \code{list} of length 4.
}
\usage{
quarter_readable
}
\description{
Quarters Readable Name
}
\keyword{datasets}
