% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingustic.R
\docType{data}
\name{RawColnameTrans}
\alias{RawColnameTrans}
\title{Translate Chinese Labels in Raw Data to English}
\format{
An object of class \code{list} of length 4.
}
\usage{
RawColnameTrans
}
\description{
Note that \code{ProductLabelName} is renamed with "product"
which is hard encoded.
}
\keyword{datasets}
