% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{scatter_echart}
\alias{scatter_echart}
\title{Scatter Chart}
\usage{
scatter_echart(
  data,
  x_var,
  y_var,
  label_var,
  x_axis_name,
  y_axis_name,
  x_lower_bound,
  x_upper_bound,
  y_lower_bound,
  y_upper_bound,
  increment = 1000,
  symbol_size = 20,
  shrink_width = NULL,
  toolbox_features = c("saveAsImage", "restore")
)
}
\arguments{
\item{data}{a data frame}

\item{x_var}{character. The variable that will be the X axis in the chart.}

\item{y_var}{character. The variable that will be the Y axis in the chart.}

\item{label_var}{character. The variable that will be the label of point in the chart.}

\item{x_axis_name}{character. The name of X axis. If missing, the function will use
information in \code{option} to auto-detect the name according to \code{x_var}.}

\item{y_axis_name}{character. The name of Y axis. If missing, the function will use
information in \code{option} to auto-detect the name according to \code{y_var}.}

\item{x_lower_bound}{numeric. The lower bound of x axis.}

\item{x_upper_bound}{numeric. The upper bound of x axis.}

\item{y_lower_bound}{numeric. The lower bound of Y axis.}

\item{y_upper_bound}{numeric. The upper bound of x axis.}

\item{increment}{numeric, default 1000. The increment of secondary Y axis ticks.}

\item{symbol_size}{numeric, default 20. The size of point in chart.}

\item{shrink_width}{character, default NULL. It uses \code{\link[echarts4r]{e_grid}} to
control the width of figure grid so the long name of X axis can be seen fully.}

\item{toolbox_features}{character vectors, default \code{c("saveAsImage", "restore")}. It controls
the toolbox functions in echarts.}
}
\value{
an object of class \code{echarts4r}.
}
\description{
Scatter Chart
}
