% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{scatter_timeline_echart}
\alias{scatter_timeline_echart}
\title{Scatter Chart with Timeline}
\usage{
scatter_timeline_echart(
  data,
  timeline_var,
  x_var,
  y_var,
  size_var,
  x_axis_name,
  y_axis_name,
  size_axis_name,
  y_lower_bound,
  y_upper_bound,
  x_label_width = 50,
  symbol_size = c(1, 25),
  timeline_top = 5,
  interval = 0,
  timeline_label_width = 50,
  toolbox_features = c("saveAsImage", "restore")
)
}
\arguments{
\item{data}{a data frame}

\item{timeline_var}{character. The variable indicating timeline.}

\item{x_var}{character. The variable that will be the X axis in the chart.}

\item{y_var}{character. The variable that will be the Y axis in the chart.}

\item{size_var}{character. The variable that will be the label of point in the chart.}

\item{x_axis_name}{character. The name of X axis. If missing, the function will use
information in \code{option} to auto-detect the name according to \code{x_var}.}

\item{y_axis_name}{character. The name of Y axis. If missing, the function will use
information in \code{option} to auto-detect the name according to \code{y_var}.}

\item{size_axis_name}{character. The name of size axis. If missing, the function will use
information in \code{option} to auto-detect the name according to \code{size_var}.}

\item{y_lower_bound}{numeric. The lower bound of Y axis.}

\item{y_upper_bound}{numeric. The upper bound of Y axis.}

\item{x_label_width}{numeric, default 50. The width of label text of x axis.}

\item{symbol_size}{numeric vector, default \code{c(1,25)}. The range of size of point in chart.}

\item{timeline_top}{numeric, default 5. The distance between timeline widget and top of chart.}

\item{interval}{numeric, default 0. The interval of label of timeline points.}

\item{timeline_label_width}{numeric, default 50. The label text width of timeline points.}

\item{toolbox_features}{character vectors, default \code{c("saveAsImage", "restore")}. It controls
the toolbox functions in echarts.}
}
\value{
an object of class \code{echarts4r}.
}
\description{
Scatter Chart with Timeline
}
