% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dimensions.R
\name{set_dimensions}
\alias{set_dimensions}
\title{Set Dimensions to be Analysis After Import antgreens}
\usage{
set_dimensions(product_label_name = NA, other_labels = NA)
}
\arguments{
\item{product_label_name}{character, default NA. It is the column name both in
raw data and MRL data. If set NA this parameter will use the default value (See
use \code{get_option("ProductLabelName")})}

\item{other_labels}{It mey be a list or a character which is one item of antgreens options.}
}
\value{
None
}
\description{
Set Dimensions to be Analysis After Import antgreens
}
\examples{
# library(antgreens)
# set_dimensions()
}
