% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_option}
\alias{set_option}
\title{Set Antgreens Option}
\usage{
set_option(var_name, value)
}
\arguments{
\item{var_name}{character. The key of variable \code{antgreens_opt_name}.}

\item{value}{New value for this option.}
}
\value{
None
}
\description{
Set Antgreens Option
}
