% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{shorten_province_name}
\alias{shorten_province_name}
\title{Shorten Province Names}
\usage{
shorten_province_name(data)
}
\arguments{
\item{data}{data frame with column named "province"}
}
\value{
a data frame which province column contains short names of province
}
\description{
Shorten Province Names
}
