% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{spec_dataset}
\alias{spec_dataset}
\title{Specify Data Set to be Used}
\usage{
spec_dataset(
  data_list,
  dims,
  ana_threshold,
  calc_type = c("product", "product_drug", "multi_residue"),
  ...,
  digits = 2
)
}
\arguments{
\item{data_list}{list of multi-dimensions combinations statistics.}

\item{dims}{character vector. All dimension variable names.}

\item{ana_threshold}{data frame of analysis threshold. This function will find
the corresponding threshold value in it.}

\item{calc_type}{character. The type of statistics to be selected which should
be in \code{c("product", "product_drug", "multi_residue")}.}

\item{...}{dimension variable names to bed selected.}

\item{digits}{numeric, default 2. Do not keep dicimal places if set NA.}
}
\value{
data frame with dimension labels and statistics.
}
\description{
Specify Data Set to be Used
}
\details{
\code{spec_dataset()} is closely related with the key names in
the return list of \code{gen_dims_comb_stats()}. This function use the same method
of variable name construction with \code{gen_dims_comb_stats()} and this method is
hard-encoding hence can not be changed.
}
