% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\docType{data}
\name{TemplateHelp}
\alias{TemplateHelp}
\title{Report Template}
\format{
a folder with three subdirectories:
\describe{
  \item{Codes}{contains \code{*.R} files which will be used in .Rmd report.}
  \item{Data}{contains raw data and processed data}
    \item{raw_data}{contains raw data files which are msotly \code{.xls(x)} files.}
  \item{Report}{bookdown project of report}
    \item{css}{forlder contains css files}
    \item{latex}{folder contains tex files}
    \item{*.Rmd}{Rmarkdown files of report}
    \item{...}{Other files bookdown needed}
}
}
\source{
\url{http://formlesslab.top/}
}
\usage{
TemplateHelp
}
\description{
This is a analysis report template of vegetables pesticide residue. You can
use this template as the beginning and develop your specified report quickly.
}
\keyword{datasets}
