% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_utils.R
\name{text_drug_detdef_max}
\alias{text_drug_detdef_max}
\title{Generate Text Describing Maximum Detection and Defective Rate for Special Item}
\usage{
text_drug_detdef_max(data, colname, specify, change = TRUE)
}
\arguments{
\item{data}{a data frame.}

\item{colname}{character. The name of the item column.}

\item{specify}{character. The specified value for \code{colname}.}

\item{change}{logical, default TRUE. Whether to change qualification rate to fraction defective.}
}
\value{
character of a sentence.
}
\description{
Generate Text Describing Maximum Detection and Defective Rate for Special Item
}
\examples{
df <- data.frame(item = c('a','b','a','a'), detection_rate_percent = runif(4,0,100), qualification_rate_percent = runif(4,0,100))
text_drug_detdef_max(df, "item", 'a')
}
