% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_utils.R
\name{text_minmax}
\alias{text_minmax}
\title{Generate Text Describing the Largest and Least Items}
\usage{
text_minmax(data, rate_text, item_name, rate_name)
}
\arguments{
\item{data}{a data frame}

\item{rate_text}{character. The character used in the text to represent the rate name.}

\item{item_name}{character. The name of column in \code{data} storing the name of item.}

\item{rate_name}{character. The name of column in \code{data} storing the value of rate.}
}
\value{
character of a sentence.
}
\description{
Generate Text Describing the Largest and Least Items
}
\examples{
df <- data.frame(rate = runif(4), item = c("a",'b','c','d'))
text_minmax(df, "\u6d4b\u8bd5\u7387", "item", "rate")
}
