% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_utils.R
\name{text_range}
\alias{text_range}
\title{Generate Range of Numeric Column in Text Form}
\usage{
text_range(data, col)
}
\arguments{
\item{data}{a data frame}

\item{col}{character. The name of a numeric column.}
}
\value{
character in the form of \code{<number><symbol> ~ <number><symbol>}
}
\description{
\code{text_range()} generate range of numeric column in data frame
in the form of \code{<number><symbol> ~ <number><symbol>}. If the column name has
"percent" string, then the \code{symbol} will be \code{'\%'}, otherwise is a void string.
}
\examples{
df <- data.frame(a = runif(10, 10, 100))
text_range(df, 'a')
}
