% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_utils.R
\name{text_threshold}
\alias{text_threshold}
\title{#' Title
#'
#' @param data
#' @param rate_text
#' @param item_name
#' @param rate_name
#'
#' @return
#' @export
#'
#' @examples
text_quarter_minmax <- function(data, rate_text, item_name, rate_name) {
  str <- paste(
    sapply(
      c("第一季度", "第二季度", "第三季度", "第四季度"),
      function(x){
        paste0(
          x,
          arrange(data %>% filter(quarter == x), desc(.data[[rate_name]]))[1,][[item_name]],
          rate_text, "最高为 ",
          arrange(data %>% filter(quarter == x), desc(.data[[rate_name]]))[1,][[rate_name]], "%，",
          arrange(data %>% filter(quarter == x), .data[[rate_name]])[1,][[item_name]],
          rate_text, "最低为 ",
          arrange(data %>% filter(quarter == x), .data[[rate_name]])[1,][[rate_name]], "%"
        )
      }
    ),
    collapse = '；')
  return(str)
}}
\usage{
text_threshold(threshold_data, dims, ana_dims)
}
\arguments{
\item{threshold_data}{a data frame with threshold information.}

\item{dims}{character vector. All dimensions.}

\item{ana_dims}{character vector. The dimensions to be analysed in this case.}
}
\value{
character of paragraph that describes the threshold in the case.
}
\description{
This function
}
\details{
#' Title
#'
#' @param data
#' @param year_range
#' @param rate_text
#' @param item_name
#' @param rate_name
#'
#' @return
#' @export
#'
#' @examples
text_timeline_minmax <- function(data, year_range, rate_text, item_name, rate_name) {
  str <- paste(
    sapply(
      year_range[1]:year_range[2],
      function(x){
        sapply(
          c("第一季度", "第二季度", "第三季度", "第四季度"),
          function(y) {
            paste0(
              x, " 年", y,
              arrange(data %>% filter(year == x, quarter == y), desc(.data[[rate_name]]))[1,][[item_name]],
              rate_text, "最高为 ",
              arrange(data %>% filter(year == x, quarter == y), desc(.data[[rate_name]]))[1,][[rate_name]], "%，",
              arrange(data %>% filter(year == x, quarter == y), .data[[rate_name]])[1,][[item_name]],
              rate_text, "最低为 ",
              arrange(data %>% filter(year == x, quarter == y), .data[[rate_name]])[1,][[rate_name]], "%"
            )
          }
        )
      }
    ),
    collapse = '；'
  )
  return(str)
}

text_year_max_3 <- function(data, year_range, rate_text, item_name, rate_name) {
  str <- paste(
    sapply(
      year_range[1]:year_range[2],
      function(x){
        paste0(
          x, " 年",
          arrange(data %>% filter(year == x), desc(.data[[rate_name]]))[1,][[item_name]],
          rate_text, "最高为 ",
          arrange(data %>% filter(year == x), desc(.data[[rate_name]]))[1,][[rate_name]], "%，",
          "其次是",
          arrange(data %>% filter(year == x), desc(.data[[rate_name]]))[2,][[item_name]], "和",
          arrange(data %>% filter(year == x), desc(.data[[rate_name]]))[3,][[item_name]],
          "，其", rate_text, "依次为 ",
          arrange(data %>% filter(year == x), desc(.data[[rate_name]]))[2,][[rate_name]], "% 和 ",
          arrange(data %>% filter(year == x), desc(.data[[rate_name]]))[3,][[rate_name]], "%"
        )
      }
    ),
    collapse = '；'
  )
  return(str)
}

text_quarter_max_3 <- function(data, rate_text, item_name, rate_name) {
  str <- paste(
    sapply(
      c("第一季度", "第二季度", "第三季度", "第四季度"),
      function(x){
        paste0(
          x,
          arrange(data %>% filter(quarter == x), desc(.data[[rate_name]]))[1,][[item_name]],
          rate_text, "最高为 ",
          arrange(data %>% filter(quarter == x), desc(.data[[rate_name]]))[1,][[rate_name]], "%，",
          "其次是",
          arrange(data %>% filter(quarter == x), desc(.data[[rate_name]]))[2,][[item_name]], "和",
          arrange(data %>% filter(quarter == x), desc(.data[[rate_name]]))[3,][[item_name]],
          "，其", rate_text, "依次为 ",
          arrange(data %>% filter(quarter == x), desc(.data[[rate_name]]))[2,][[rate_name]], "% 和 ",
          arrange(data %>% filter(quarter == x), desc(.data[[rate_name]]))[3,][[rate_name]], "%"
        )
      }
    ),
    collapse = '；'
  )
  return(str)
}

text_timeline_max_3 <- function(data, year_range, rate_text, item_name, rate_name) {
  str <- paste(
    sapply(
      year_range[1]:year_range[2],
      function(x){
        sapply(
          c("第一季度", "第二季度", "第三季度", "第四季度"),
          function(y) {
            paste0(
              x, " 年", y,
              arrange(data %>% filter(year == x, quarter == y), desc(.data[[rate_name]]))[1,][[item_name]],
              rate_text, "最高为 ",
              arrange(data %>% filter(year == x, quarter == y), desc(.data[[rate_name]]))[1,][[rate_name]], "%，",
              "其次是",
              arrange(data %>% filter(year == x, quarter == y), desc(.data[[rate_name]]))[2,][[item_name]], "和",
              arrange(data %>% filter(year == x, quarter == y), desc(.data[[rate_name]]))[3,][[item_name]],
              "，其", rate_text, "依次为 ",
              arrange(data %>% filter(year == x, quarter == y), desc(.data[[rate_name]]))[2,][[rate_name]], "% 和 ",
              arrange(data %>% filter(year == x, quarter == y), desc(.data[[rate_name]]))[3,][[rate_name]], "%"
            )
          }
        )
      }
    ),
    collapse = '；'
  )
  return(str)
}


#' Title
#'
#' @param data
#' @param year_range
#' @param rate_text
#' @param item_name_1
#' @param item_name_2
#' @param rate_name
#'
#' @return
#' @export
#'
#' @examples
text_timeline_two_max <- function(
  data, year_range, rate_text, item_name_1, item_name_2, rate_name
) {
  str <- paste(
    sapply(
      year_range[1]:year_range[2],
      function(x){
        sapply(
          c("第一季度", "第二季度", "第三季度", "第四季度"),
          function(y) {
            paste0(
              x, " 年", y,
              arrange(data %>% filter(year == x, quarter == y), desc(.data[[rate_name]]))[1,][[item_name_1]],
              arrange(data %>% filter(year == x, quarter == y), desc(.data[[rate_name]]))[1,][[item_name_2]],
              rate_text, "最高为 ",
              arrange(data %>% filter(year == x, quarter == y), desc(.data[[rate_name]]))[1,][[rate_name]], "%"
            )
          }
        )
      }
    ),
    collapse = '；'
  )
  return(str)
}

text_year_two_max_3 <- function(data, year_range, rate_text, item_name_1, item_name_2, rate_name) {
  str <- paste(
    sapply(
      year_range[1]:year_range[2],
      function(x){
        paste0(
          x, " 年",
          arrange(data %>% filter(year == x), desc(.data[[rate_name]]))[1,][[item_name_1]], "的",
          arrange(data %>% filter(year == x), desc(.data[[rate_name]]))[1,][[item_name_2]],
          rate_text, "最高为 ",
          arrange(data %>% filter(year == x), desc(.data[[rate_name]]))[1,][[rate_name]], "%，",
          "其次是",
          arrange(data %>% filter(year == x), desc(.data[[rate_name]]))[2,][[item_name_1]], "的",
          arrange(data %>% filter(year == x), desc(.data[[rate_name]]))[2,][[item_name_2]],"和",
          arrange(data %>% filter(year == x), desc(.data[[rate_name]]))[3,][[item_name_1]], "的",
          arrange(data %>% filter(year == x), desc(.data[[rate_name]]))[3,][[item_name_2]],
          "，其", rate_text, "依次为 ",
          arrange(data %>% filter(year == x), desc(.data[[rate_name]]))[2,][[rate_name]], "% 和 ",
          arrange(data %>% filter(year == x), desc(.data[[rate_name]]))[3,][[rate_name]], "%"
        )
      }
    ),
    collapse = '；'
  )
  return(str)
}

text_category_max_3 <- function(data, rate_text, item_name, rate_name) {
  category_names <- unique(data[["category"]])
  str <- paste(
    sapply(
      category_names,
      function(x){
        paste0(
          x, "在",
          arrange(data %>% filter(category == x), desc(.data[[rate_name]]))[1,][[item_name]], "的",
          rate_text, "最高为 ",
          arrange(data %>% filter(category == x), desc(.data[[rate_name]]))[1,][[rate_name]], "%，",
          "其次是",
          arrange(data %>% filter(category == x), desc(.data[[rate_name]]))[2,][[item_name]], "和",
          arrange(data %>% filter(category == x), desc(.data[[rate_name]]))[3,][[item_name]],
          "，其", rate_text, "依次为 ",
          arrange(data %>% filter(category == x), desc(.data[[rate_name]]))[2,][[rate_name]], "% 和 ",
          arrange(data %>% filter(category == x), desc(.data[[rate_name]]))[3,][[rate_name]], "%"
        )
      }
    ),
    collapse = '；'
  )
  return(str)
}


#' Title
#'
#' @param data
#' @param rate_text
#' @param item_name
#' @param rate_name
#'
#' @return
#' @export
#'
#' @examples
text_category_minmax <- function(data, rate_text, item_name, rate_name) {
  category_names <- unique(data[["category"]])
  str <- paste(
    sapply(
      category_names,
      function(x){
        paste0(
          x, "在",
          arrange(data %>% filter(category == x), desc(.data[[rate_name]]))[1,][[item_name]],
          rate_text, "最高为 ",
          arrange(data %>% filter(category == x), desc(.data[[rate_name]]))[1,][[rate_name]], "%，",
          arrange(data %>% filter(category == x), .data[[rate_name]])[1,][[item_name]],
          rate_text, "最低为 ",
          arrange(data %>% filter(category == x), .data[[rate_name]])[1,][[rate_name]], "%"
        )
      }
    ),
    collapse = '；')
  return(str)
}

#' Title
#'
#' @param data
#' @param colname
#' @param specify
#' @param change
#'
#' @return
#' @export
#'
#' @examples
text_drug_detdef_max <- function(data, colname, specify, change = TRUE) {
  if (change) data <- data %>% change_to_defective()
  str <- paste0(
    specify, "在",
    arrange(data %>% filter(.data[[colname]] == specify), desc(detection_rate_percent))[["drug"]][1],
    "上的检出率最高为 ",
    arrange(data %>% filter(.data[[colname]] == specify), desc(detection_rate_percent))[["detection_rate_percent"]][1],
    "%，在",
    arrange(data %>% filter(.data[[colname]] == specify), desc(defective_rate_percent))[["drug"]][1],
    "上的超标率最高为 ",
    arrange(data %>% filter(.data[[colname]] == specify), desc(defective_rate_percent))[["defective_rate_percent"]][1],
    "%"
  )
  return(str)
}
Generate Text Describing the Threshold
}
\examples{
df <- data.frame(year=c(1,0,0,1),quarter=c(0,1,0,0),province=c(0,0,1,1),threshold = runif(4))
text_threshold(df, c('year','quarter','province'), c('year', 'province'))
}
