% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_utils.R
\name{text_timeline_two_max}
\alias{text_timeline_two_max}
\title{Generate Text Describing Three Maximum Two-Items By Year-Quarter}
\usage{
text_timeline_two_max(
  data,
  year_range,
  rate_text,
  item_name_1,
  item_name_2,
  rate_name
)
}
\arguments{
\item{data}{a data frame}

\item{year_range}{numeric vector represents the year.}

\item{rate_text}{character. The character used in the text to represent the rate name.}

\item{item_name_1}{character. The name of column in \code{data} storing the name of first item.}

\item{item_name_2}{character. The name of column in \code{data} storing the name of second item.}

\item{rate_name}{character. The name of column in \code{data} storing the value of rate.}
}
\value{
a character of sentence.
}
\description{
Generate Text Describing Three Maximum Two-Items By Year-Quarter
}
