% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_utils.R
\name{text_two_max_3}
\alias{text_two_max_3}
\title{Generate Text Describing Three Largest Two-Items}
\usage{
text_two_max_3(
  data,
  rate_text,
  item_name_1,
  item_name_2,
  rate_name,
  conjunction = "中的"
)
}
\arguments{
\item{data}{a data frame}

\item{rate_text}{character. The character used in the text to represent the rate name.}

\item{item_name_1}{character. The name of column in \code{data} storing the name of first item.}

\item{item_name_2}{character. The name of column in \code{data} storing the name of second item.}

\item{rate_name}{character. The name of column in \code{data} storing the value of rate.}

\item{conjunction}{character. It is the conjunction between the first
item and the second item to make the sentence more fluent.}
}
\value{
character of a sentence.
}
\description{
Generate Text Describing Three Largest Two-Items
}
\examples{
df <- data.frame(rate = runif(4), item1 = c("a",'b','c','d'), item2 = c("ab",'bc','cd','de'))
text_two_max_3(df, "\u6d4b\u8bd5\u7387", "item1", "item2", "rate")
text_two_max_3(df, "\u6d4b\u8bd5\u7387", "item1", "item2", "rate", conjunction = "")
}
