% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_utils.R
\name{text_year_minmax}
\alias{text_year_minmax}
\title{Generate Text Describing Minimum and Maximum Rate Item By Year}
\usage{
text_year_minmax(data, year_range, rate_text, item_name, rate_name)
}
\arguments{
\item{data}{a data frame}

\item{year_range}{numeric vector represents the year.}

\item{rate_text}{character. The character used in the text to represent the rate name.}

\item{item_name}{character. The name of column in \code{data} storing the name of item.}

\item{rate_name}{character. The name of column in \code{data} storing the value of rate.}
}
\value{
a character of sentence.
}
\description{
Generate text describing minimum and maximum rate item for each year.
}
