% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_utils.R
\name{text_year_two_max_3}
\alias{text_year_two_max_3}
\title{Generate Text Describing Three Maximum Two-Items By Year}
\usage{
text_year_two_max_3(
  data,
  year_range,
  rate_text,
  item_name_1,
  item_name_2,
  rate_name,
  conjunction = "的"
)
}
\arguments{
\item{data}{a data frame}

\item{year_range}{numeric vector represents the year.}

\item{rate_text}{character. The character used in the text to represent the rate name.}

\item{item_name_1}{character. The name of column in \code{data} storing the name of first item.}

\item{item_name_2}{character. The name of column in \code{data} storing the name of second item.}

\item{rate_name}{character. The name of column in \code{data} storing the value of rate.}

\item{conjunction}{character. It is the conjunction between the first
item and the second item to make the sentence more fluent.}
}
\value{
a character of sentence.
}
\description{
Generate Text Describing Three Maximum Two-Items By Year
}
