% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingustic.R
\docType{data}
\name{ThresholdName}
\alias{ThresholdName}
\title{Name of Threshold Column}
\format{
An object of class \code{character} of length 1.
}
\usage{
ThresholdName
}
\description{
Name of Threshold Column
}
\keyword{datasets}
