% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_report.R
\name{update_params}
\alias{update_params}
\title{Update Compile Parameters}
\usage{
update_params(new_params_list)
}
\arguments{
\item{new_params_list}{list. The given parameters.}
}
\value{
list.
}
\description{
Use given compile parameters to update the default parameters.
}
