% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{use_template}
\alias{use_template}
\title{Copy Report Template to Directory}
\usage{
use_template(proj_dir, proj_name = "vegetables_residue_report")
}
\arguments{
\item{proj_dir}{path of report directory.}

\item{proj_name}{name of directory which will be the report project.}
}
\value{
None
}
\description{
Copy Report Template to Directory
}
