% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{arrange_rows}
\alias{arrange_rows}
\title{Arrange Rows by Multiple Columns}
\usage{
arrange_rows(data, ...)
}
\arguments{
\item{data}{a data frame}

\item{...}{quoted column names}
}
\value{
a data frame with arranged rows
}
\description{
Arrange rows in a data frame according by the columns provided. The order of
columns will determine the order of arrangement.
}
\examples{
df <- data.frame(a = c(4, 6, 2), b = c('a', 's', 'd'))
df <- arrange_rows(df, a, b)
}
