% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_report.R
\name{book_env}
\alias{book_env}
\title{Generate Environment for Book Rendering}
\usage{
book_env(params = list(), year_colname, year_range = NA)
}
\arguments{
\item{params}{list. The parameters of report.}

\item{year_colname}{character of column name for dimension \code{year}.}

\item{year_range}{numeric vector length 2, default NA.}
}
\value{
environment object contains necessary objects.
}
\description{
Generate Environment for Book Rendering
}
