% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{factor_dims}
\alias{factor_dims}
\title{Make Columns As Factor}
\usage{
factor_dims(dataset, dim_names, funcs = list(), quarter_readable = TRUE)
}
\arguments{
\item{dataset}{data frame with readable "quarter" column
or "year"/"province"/"category"/"product"/"drug" column.}

\item{dim_names}{character vector.}

\item{funcs}{named list of customed factor function}

\item{quarter_readable}{logical, default TRUE.}
}
\value{
data frame with factor column. And the order of quarter is
depends on the order of \code{quarter_readable}; the order of year is depends on
the number of year; the order of province, category product and drug are depends
on the dictionary order of character.
}
\description{
Make Columns As Factor
}
