% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_utils.R
\name{text_threshold}
\alias{text_threshold}
\title{Generate Text Describing the Threshold}
\usage{
text_threshold(threshold_data, dims, ana_dims)
}
\arguments{
\item{threshold_data}{a data frame with threshold information.}

\item{dims}{character vector. All dimensions.}

\item{ana_dims}{character vector. The dimensions to be analysed in this case.}
}
\value{
character of paragraph that describes the threshold in the case.
}
\description{
This function
}
\examples{
df <- data.frame(year=c(1,0,0,1),quarter=c(0,1,0,0),province=c(0,0,1,1),threshold = runif(4))
text_threshold(df, c('year','quarter','province'), c('year', 'province'))
}
